[domain] create api.controller package & move controller class
